fix(tests): update import paths after component restructure
Updated test file imports to reflect the new modular component directory structure:
- photo-viewer.test.tsx: mock paths for photo-metadata-panel and live-photo-player
- live-photo-player.test.tsx: import from @/components/shared/live-photo-player
- photo-metadata-panel.test.tsx: import from ../photo-viewer/photo-metadata-panel
- auth-header.test.tsx: import from ../auth/auth-header
- email-login-form.test.tsx: import from ../auth/email-login-form
- gallery-card.test.tsx: import from ../gallery/gallery-card and mock path fix
- language-switcher.test.tsx: import from ../layout/language-switcher
- like-button.test.tsx: import from ../shared/like-button
