+++ src/components/__tests__/auth-header.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
-import { AuthHeader } from "../auth-header";
+import { AuthHeader } from "../auth/auth-header";
 
 const pushMock = vi.fn();
