+++ src/components/__tests__/photo-viewer.test.tsx
@@ -30,20 +30,20 @@ vi.mock("next/link", () => ({
 
 // Mock Next.js Image
 vi.mock("next/image", () => ({
-  default: ({ src, alt, ...props }: { src: string; alt: string; [key: string]: unknown }) => (
+  default: ({ src, alt, ...props }: { src: string; alt: string;[key: string]: unknown }) => (
     <img src={src} alt={alt} {...props} />
   ),
 }));
 
 // Mock PhotoMetadataPanel
-vi.mock("../photo-metadata-panel", () => ({
+vi.mock("../photo-viewer/photo-metadata-panel", () => ({
   PhotoMetadataPanel: ({ image }: { image: { title: string } }) => (
     <div data-testid="metadata-panel">Metadata for {image.title}</div>
   ),
 }));
 
 // Mock LivePhotoPlayer
-vi.mock("../live-photo-player", () => ({
+vi.mock("@/components/shared/live-photo-player", () => ({
   LivePhotoPlayer: ({
     imageSrc,
     videoSrc,
@@ -67,9 +67,9 @@ class MockXHR {
   onload: ((this: XMLHttpRequest, ev: Event) => unknown) | null = null;
   onprogress:
     | ((
-        this: XMLHttpRequest,
-        ev: Event & { loaded?: number; total?: number; lengthComputable?: boolean }
-      ) => unknown)
+      this: XMLHttpRequest,
+      ev: Event & { loaded?: number; total?: number; lengthComputable?: boolean }
+    ) => unknown)
     | null = null;
   onerror: ((this: XMLHttpRequest, ev: Event) => unknown) | null = null;
   onabort: ((this: XMLHttpRequest, ev: Event) => unknown) | null = null;
